# docs/swagger.yaml
definitions:
  auth.AccessTokenRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  auth.FetchUserResponse:
    properties:
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  permission.CreatePermissionRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: integer
    type: object
  permission.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  permission.ListPermissionsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/permission.PermissionResponse'
        type: array
    type: object
  permission.PermissionResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      updated_at:
        type: string
    type: object
  permission.UpdatePermissionRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /auth/access-token:
    post:
      consumes:
      - application/json
      description: Authenticate using email and password
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Authenticate a user and receive access/refresh tokens
      tags:
      - Auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Refresh a user's access token
      tags:
      - Auth
  /auth/user:
    get:
      consumes:
      - application/json
      description: Fetch logged-in user details from access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.FetchUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - Auth
  /permissions/:
    get:
      consumes:
      - application/json
      description: Get a paginated list of permissions
      parameters:
      - description: Maximum results
        in: query
        name: limit
        type: integer
      - description: Result offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.ListPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      description: Create a permission with name, code, optional description, and
        optional parent permission
      parameters:
      - description: Create permission request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/permission.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/permission.PermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - Permissions
  /permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a permission
      tags:
      - Permissions
    get:
      consumes:
      - application/json
      description: Get a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.PermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get permission details
      tags:
      - Permissions
    put:
      consumes:
      - application/json
      description: Update permission fields by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update permission request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/permission.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.PermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a permission
      tags:
      - Permissions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"